requirements.txt
streamlit>=1.30
google-genai>=1.0
Pillow>=10.0
aiogram>=3.10
